feat(textarea): findWordRight / findWordLeft
